rename key_switcheroo to just switcheroo
