+++ switcheroo/util.py
